DEVEL15-linux-warning-reduction-20090318
LICENSE IPL10
FIXES 119434

cleanup warnings in linux kernel builds


(cherry picked from commit 85d8bdac7439966a64709724d20dfd487911f7bc)

+++ src/afs/afs_osi_alloc.c
@@ -168,7 +168,7 @@ osi_AllocLargeSpace(size_t size)
 
     AFS_STATCNT(osi_AllocLargeSpace);
     if (size > AFS_LRALLOCSIZ)
-	osi_Panic("osi_AllocLargeSpace: size=%d\n", size);
+	osi_Panic("osi_AllocLargeSpace: size=%d\n", (int)size);
     afs_stats_cmperf.LargeBlocksActive++;
     if (!freePacketList) {
 	char *p;
@@ -202,7 +202,7 @@ osi_AllocSmallSpace(size_t size)
 
     AFS_STATCNT(osi_AllocSmallSpace);
     if (size > AFS_SMALLOCSIZ)
-	osi_Panic("osi_AllocSmallS: size=%d\n", size);
+	osi_Panic("osi_AllocSmallS: size=%d\n", (int)size);
 
     if (!freeSmallList) {
 	afs_stats_cmperf.SmallBlocksAlloced++;
